Move api request to shared lib
