+++ app/frontend-webapp/api/studies.ts
@@ -4,8 +4,8 @@ import { revalidatePath } from "next/cache";
 
 import { Study, StudyPartial } from "@/types/studies";
 
-import { ApiError } from "./error";
-import request from "./request";
+import { ApiError } from "../lib/api/error";
+import request from "../lib/api/request";
 
 const fetchKeys = {
   list: "studies",
